# CITFAC-openapi.yaml
openapi: 3.0.3
info:
  title: Città Facile
  description: Insieme di API esposte dal BFF di Città Facile (nativamente o attraverso proxy di API esistenti)
  contact:
    email: maurizio.peisino@csi.it
  version: 1.0.0
servers:
  - url: 'https://localhost'
...
paths:
  '/api-admin/tenant/{tenantId}/cms':
    get:
      operationId: getCms
      summary: Get cms
      description: Get cms
      tags:
        - Administration | Cms
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: test
          content:
            application/json:
              schema:
                type: object
                properties:
                  backOffice:
                    type: string
                    description: Back Office
                    example: Sample Back Office
                  urlServizi:
                    type: string
                    description: URL for Servizi
                    example: 'http://sample.com/servizi'
                  urlMenu:
                    type: string
                    description: URL for Menu
                    example: 'http://sample.com/menu'
        '404':
          $ref: '#/components/responses/404'
    post:
      operationId: postCms
      summary: Generate Cms
      description: Create a new Cms object
      tags:
        - Administration | Cms
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                backOffice:
                  type: string
                  description: Back Office
                  example: Sample Back Office
                urlServizi:
                  type: string
                  description: URL for Servizi
                  example: 'http://sample.com/servizi'
                urlMenu:
                  type: string
                  description: URL for Menu
                  example: 'http://sample.com/menu'
                comune:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: Comune ID
                      example: 1
                    name:
                      type: string
                      description: Comune Name
                      example: Sample Comune
              required:
                - backOffice
                - urlServizi
                - urlMenu
                - comune
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/400'
    put:
      operationId: putCms
      summary: Update cms
      description: Update cms
      tags:
        - Administration | Cms
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                backOffice:
                  type: string
                  description: Back Office
                  example: Sample Back Office
                urlServizi:
                  type: string
                  description: URL for Servizi
                  example: 'http://sample.com/servizi'
                urlMenu:
                  type: string
                  description: URL for Menu
                  example: 'http://sample.com/menu'
                comune:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: Comune ID
                      example: 1
                    name:
                      type: string
                      description: Comune Name
                      example: Sample Comune
              required:
                - backOffice
                - urlServizi
                - urlMenu
                - comune
      responses:
        '200':
          description: Cms Updated
        '400':
          $ref: '#/components/responses/400'
    delete:
      operationId: removeCms
      summary: Delete Cms
      description: Delete a cms
      tags:
        - Administration | Cms
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cms deleted
        '400':
          $ref: '#/components/responses/400'
...
  '/api-admin/tenant/{tenantId}/feedback':
    get:
      operationId: getFeedback
      summary: Get Feedbacks
      description: get feedback  by using path params
      tags:
        - Administration | Feedback
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
        - in: query
          name: serviceId
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
        - in: query
          name: to
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
  '/api-admin/tenant/{tenantId}/feedback/messages':
    get:
      operationId: getMessages
      summary: Get Messages
      description: get feedback messages by using path params
      tags:
        - Administration | Feedback
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: integer
        - in: query
          name: serviceId
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
        - in: query
          name: to
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: startPage
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
...
  /api-admin/tenant:
    post:
      operationId: addTenant
      summary: Add tenant
      description: Add new tenant configuration
      tags:
        - Administration | Tenants
      responses:
        '201':
          description: Create tenant configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComuneAdmin'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '422':
          $ref: '#/components/responses/422'
  '/api-admin/tenant/{tenant-id}':
    parameters:
      - $ref: '#/components/parameters/idComune'
    get:
      operationId: getTenant
      summary: Get tenant
      description: Get single tenant configuration
      tags:
        - Administration | Tenants
      responses:
        '200':
          description: Tenant configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComuneAdmin'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
    put:
      operationId: putTenant
      summary: Update tenant
      description: Update single tenant configuration
      tags:
        - Administration | Tenants
      responses:
        '200':
          description: Updated tenant configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComuneAdmin'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
    patch:
      operationId: patchTenant
      summary: Update tenant partially
      description: Update single tenant configuration passing modified fields only
      tags:
        - Administration | Tenants
      responses:
        '200':
          description: Updated tenant configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComuneAdmin'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
  '/api-admin/tenant/{tenant-id}/services':
    parameters:
      - $ref: '#/components/parameters/idComune'
    get:
      operationId: getServices
      summary: Get all services
      description: Get all services for single tenant
      tags:
        - Administration | Services
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Servizio'
    delete:
      operationId: deleteServices
      summary: Delete all services
      description: Delete all services for single tenant
      tags:
        - Administration | Services
      responses:
        '200':
          description: Services list is deleted
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
  '/api-admin/tenant/{tenant-id}/service':
    parameters:
      - $ref: '#/components/parameters/idComune'
    post:
      operationId: addService
      summary: Add service
      description: Add new service for single tenant
      tags:
        - Administration | Services
      responses:
        '201':
          description: Create tenant service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Servizio'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
  '/api-admin/tenant/{tenant-id}/service/{service-id}':
    parameters:
      - $ref: '#/components/parameters/idComune'
      - $ref: '#/components/parameters/idServizio'
    get:
      operationId: getService
      summary: Get service
      description: Get single service for single tenant
      tags:
        - Administration | Services
      responses:
        '200':
          description: Tenant service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Servizio'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
    put:
      operationId: putService
      summary: Update service
      description: Update single service for single tenant
      tags:
        - Administration | Services
      responses:
        '200':
          description: Updated tenant service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Servizio'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
    delete:
      operationId: deleteService
      summary: Delete service
      description: Delete single service for single tenant
      tags:
        - Administration | Services
      responses:
        '200':
          description: Service is deleted
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
  '/api-admin/tenant/{tenant-id}/details':
    parameters:
      - $ref: '#/components/parameters/idComune'
    get:
      operationId: getDetails
      summary: Get all details
      description: Get all services details for single tenant
      tags:
        - Administration | Services details
      responses:
        '200':
          description: List of services details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DettaglioServizio'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
  '/api-admin/tenant/{tenant-id}/detail':
    parameters:
      - $ref: '#/components/parameters/idComune'
    post:
      operationId: addDetail
      summary: Add detail
      description: Add new service detail for single tenant
      tags:
        - Administration | Services details
      responses:
        '201':
          description: Create tenant service detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DettaglioServizio'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
...
  '/api-admin/tenant/{tenant-id}/credentials':
    parameters:
      - $ref: '#/components/parameters/idComune'
    get:
      operationId: getCredentials
      summary: Get all credentials
      description: Get all credentials for single tenant
      tags:
        - Administration | Credentials
      responses:
        '200':
          description: List of credentials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Credenziale'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
...
components:
  parameters:
...
    idServizio:
      name: service-id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    '400':
      description: Bad Request. Body or parameters validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
...
    '404':
      description: Not Found. Cannot find entity key or tenant id related to the entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
...
  schemas:
...
    ComuneAdmin:
      allOf:
        - $ref: '#/components/schemas/Comune'
        - type: object
          properties:
            id:
              type: integer
              example: 1
            templateNotificatore:
              type: string
            logoutUri:
              type: string
            cf:
              type: string
            cms:
              type: object
              properties:
                urlServizi:
                  type: string
                urlMenu:
                  type: string
                backOffice:
                  type: string
            assets:
              type: array
              items:
                type: object
                properties:
                  tipoCredenziali:
                    type: string
                  urlBackoffice:
                    type: string
                  icon:
                    type: string
            apiManagerKey:
              type: string
            apiManagerSecret:
              type: string
...